Fixed clearGitFiles and refreshGithubRepo

+++ DuggaSys/microservices/gitCommitService/clearGitFiles_ms.php
@@ -6,19 +6,25 @@ include_once "../../../Shared/sessions.php";
 
 global $pdo;
 
+header("Content-Type: application/json");
+
 //--------------------------------------------------------------------------------------------------
 // clearGitFiles: Clear the gitFiles table in SQLite db when a course has been updated with a new github repo. This function is used by other github microservices.
 //--------------------------------------------------------------------------------------------------
 
-function clearGitFiles($cid) {
+if ($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_POST['cid'])) {
+    $cid = $_POST['cid'];
+    
     $pdolite = new PDO('sqlite:../../githubMetadata/metadata2.db');
     $query = $pdolite->prepare("DELETE FROM gitFiles WHERE cid = :cid"); 
     $query->bindParam(':cid', $cid);
+    
     if (!$query->execute()) {
         $error = $query->errorInfo();
-        echo "Error updating file entries" . $error[2];
-        $errorvar = $error[2];
-        print_r($error);
-        echo $errorvar;
+        echo json_encode(["success" => false, "error" => $error[2]]);
+        exit;
     }
 }
+
+// Returns JSON
+echo json_encode(["success" => true]);
